test(util): migrating unit tests to TypeScript
This involves a number of changes:

- Added a "production" tsconfig file called tsconfig.prod.json. This file contains settings specific for production builds of the package, but extends the existing tsconfig.json for easier maintenance.
- The existing tsconfig.json file has been tweaked to be more iteration- and developer-friendly. It adds mocha types (so "describe", "it", and similar are recognized in ts spec files), and omits some strict code checks (unused locals, for examples). These have been moved to the production config instead.
- Changed build and test scripts in package.json to target new tsconfig file and TS unit tests.
- Migrated one unit test to typescript as an example.
- DOWNGRADED chai to v4 instead of v5. Chai v5 requires use of ESM and the current codebase still compiles to CJS. I think the change would be relatively easy, but I don't want to do too many things in these commits.
