+++ packages/util/.mocharc.js
@@ -1,7 +1,10 @@
 module.exports = {
   "color": true,
   "ui": "bdd",
-  "spec": "test/**/*.spec.mjs",
+  "spec": "test/**/*.spec.ts",
+  "require": [
+    "ts-node/register",
+  ],
   // Uncomment to get a JSON file with test results instead of output in console.
   // "reporter": "json",
   // "reporterOptions": [
