+++ packages/util/package.json
@@ -8,7 +8,7 @@
   ],
   "scripts": {
     "tsc": "tsc",
-    "test": "mocha --require esm",
+    "test": "mocha --spec=test/**/*.spec.ts",
     "test:ANameTest": "mocha --require esm test/ANameTest.spec.mjs",
     "test:ANameTestWiki": "mocha --require esm test/ANameTestWiki.spec.js",
     "test:svg": "mocha --require esm test/ANameSVGTest.spec.js",
@@ -16,7 +16,7 @@
     "lint": "eslint src --ext .ts",
     "lint:fix": "eslint src --ext .ts --fix",
     "format": "prettier src/**/*.ts --write",
-    "prebuild": "rimraf prebuild && tsc --build",
+    "prebuild": "rimraf prebuild && tsc --build tsconfig.prod.json",
     "build": "rimraf dist && npm run prebuild && webpack --mode production && rimraf prebuild",
     "docs": "typedoc --out docs src/index.ts",
     "prepublishOnly": "npm run build"
@@ -42,9 +42,11 @@
     "@qualweb/qw-page": "^0.2.19",
     "@qualweb/types": "0.7.26",
     "@tsconfig/recommended": "^1.0.3",
+    "@types/chai": "^4.3.11",
+    "@types/mocha": "^10.0.6",
     "@types/node": "^15.6.1",
     "@types/string-pixel-width": "^1.10.3",
-    "chai": "^5.0.0",
+    "chai": "^4.4.1",
     "css-select": "^5.1.0",
     "eslint": "^8.56.0",
     "esm": "^3.2.25",
@@ -52,8 +54,9 @@
     "prettier": "^3.1.1",
     "puppeteer": "^21.6.1",
     "rimraf": "^5.0.5",
+    "ts-node": "^10.9.2",
     "typedoc": "^0.25.6",
-    "typescript": "^4.2.4",
+    "typescript": "5.3.2",
     "webpack": "^5.89.0",
     "webpack-cli": "^4.7.0"
   },
