+++ packages/util/tsconfig.json
@@ -4,7 +4,7 @@
     /* Basic Options */
     "resolveJsonModule": true,
     "outDir": "./prebuild" /* Redirect output structure to the directory. */,
-    "rootDir": "./src" /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */,
+    "rootDir": "." /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */,
     "removeComments": true /* Do not emit comments to output. */,
 
     /* Strict Type-Checking Options */
@@ -16,19 +16,20 @@
     "noImplicitThis": true /* Raise error on 'this' expressions with an implied 'any' type. */,
     "alwaysStrict": true /* Parse in strict mode and emit "use strict" for each source file. */,
 
-    /* Additional Checks */
-    "noUnusedLocals": true /* Report errors on unused locals. */,
-    "noUnusedParameters": true /* Report errors on unused parameters. */,
-    "noImplicitReturns": true /* Report error when not all code paths in function return a value. */,
-    "noFallthroughCasesInSwitch": true /* Report errors for fallthrough cases in switch statement. */,
-
     /* Module Resolution Options */
     "moduleResolution": "node" /* Specify module resolution strategy: 'node' (Node.js) or 'classic' (TypeScript pre-1.6). */,
-    "typeRoots": ["@qualweb/types"] /* List of folders to include type definitions from. */,
-    "types": ["@qualweb/types", "node", "string-pixel-width"] /* Type declaration files to be included in compilation. */,
+    "types": [
+      "node",
+      "string-pixel-width",
+      "mocha",
+    ] /* Type declaration files to be included in compilation. */,
 
     /* Experimental Options */
     "experimentalDecorators": true /* Enables experimental support for ES7 decorators. */,
-    "emitDecoratorMetadata": true /* Enables experimental support for emitting type metadata for decorators. */
-  }
+    "emitDecoratorMetadata": true /* Enables experimental support for emitting type metadata for decorators. */,
+  },
+  "include": [
+    "src/**/*.ts",
+    "test/**/*.ts",
+  ]
 }
